Parola suplimentara pentru submit/delete

+++ delete_table.php
@@ -1,7 +1,13 @@
 <?php
 	
 	$tableNumber = $_GET["table"];
-	$mysqli = new mysqli("localhost", "php", "php", "datadb");
-	$result = $mysqli->query("DELETE FROM table".$tableNumber);
-	
+    $pass_get = $_GET["pass"];
+	include('Backend/config.php');
+    if($pass_get == $pass_script){
+        $mysqli = new mysqli($host, $user, $pass, $database);
+        $result = $mysqli->query("DELETE FROM table".$tableNumber);
+        echo "OK";
+	} else {
+        echo "Parola gresita, boss";
+    }
 ?>
